+++ packages/maas/frontend/src/app/utilities/phaseLabelUtils.tsx
@@ -1,12 +1,7 @@
 import * as React from 'react';
 import {
   AlertProps,
-  ClipboardCopyButton,
-  CodeBlockAction,
-  CodeBlockCode,
   ContentProps,
-  CodeBlock,
-  ExpandableSection,
   Icon,
   LabelProps,
   Stack,
@@ -19,6 +14,7 @@ import {
   ExclamationTriangleIcon,
   PendingIcon,
 } from '@patternfly/react-icons';
+import PhaseApiDetails from '~/app/shared/PhaseApiDetails';
 
 type PopoverContent = {
   headerIcon: React.ReactNode;
@@ -270,7 +266,7 @@ export const getModalSubtitle = (resourceType: PhaseResourceType): string | unde
 export const getModalAlertProps = (
   phase: string,
   resourceType: PhaseResourceType,
-  statusMessage?: React.ReactNode,
+  statusMessage?: string,
   reason?: string,
   status?: string,
   type?: string,
@@ -311,67 +307,9 @@ export const getModalAlertProps = (
   };
 };
 
-type PhaseApiDetailsProps = {
-  reason?: string;
-  statusMessage?: React.ReactNode;
-  status?: string;
-  type?: string;
-  lastTransitionTime?: string;
-};
-
-const PhaseApiDetails: React.FC<PhaseApiDetailsProps> = ({
-  reason,
-  statusMessage,
-  status,
-  type,
-  lastTransitionTime,
-}) => {
-  const [isExpanded, setIsExpanded] = React.useState(false);
-  const [copied, setCopied] = React.useState(false);
-  const clipboardCopyFunc = (text: string) => {
-    navigator.clipboard.writeText(text.toString());
-  };
-
-  const onClick = (text: string) => {
-    clipboardCopyFunc(text);
-    setCopied(true);
-  };
-
-  const json = convertStatusDetailsToJson(reason, statusMessage, status, type, lastTransitionTime);
-
-  const detailsActions = (
-    <CodeBlockAction>
-      <ClipboardCopyButton
-        id="status-details-copy-button"
-        aria-label="Copy to clipboard status details"
-        onClick={() => onClick(json)}
-        exitDelay={copied ? 1500 : 600}
-        maxWidth="110px"
-        variant="plain"
-        onTooltipHidden={() => setCopied(false)}
-      >
-        {copied ? 'Successfully copied to clipboard!' : 'Copy to clipboard'}
-      </ClipboardCopyButton>
-    </CodeBlockAction>
-  );
-
-  return (
-    <ExpandableSection
-      toggleText={isExpanded ? 'Hide API details' : 'API details'}
-      isExpanded={isExpanded}
-      onToggle={(_event, expanded) => setIsExpanded(expanded)}
-      data-testid="phase-api-details"
-    >
-      <CodeBlock actions={detailsActions}>
-        <CodeBlockCode data-testid="phase-api-details-code-block">{json}</CodeBlockCode>
-      </CodeBlock>
-    </ExpandableSection>
-  );
-};
-
-const convertStatusDetailsToJson = (
+export const convertStatusDetailsToJson = (
   reason?: string,
-  statusMessage?: React.ReactNode,
+  statusMessage?: string,
   status?: string,
   type?: string,
   lastTransitionTime?: string,
@@ -520,6 +458,11 @@ export const getSubtextProps = (phase: string): ContentProps | undefined => {
     case PhaseStatus.FAILED:
       return {
         className: 'pf-v6-u-text-color-status-danger',
+        style: sharedStyle,
+      };
+    case PhaseStatus.PENDING:
+      return {
+        className: 'pf-v6-u-text-color-status-info',
         style: sharedStyle,
       };
     default:
